+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "The dart 101".
 - Tests for "Survey prediction".
 - Tests for "Flip the sign".
+- Tests for "The michelangelo code".
 
 ## [1.9.0] - 2022-06-01
 ### Added